auth_api/requirements.txt
pymongo
fastapi
PyJWT